requirements.txt
Pillow==10.0.0
requests==2.31.0
pyinstaller==6.0.0